legacy: move UpgraderFrom12To20 to borg.legacy, refs #9556
